requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-environ==0.8.1
djangorestframework==3.13.0
env==0.1.0
environ==1.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
